More API Changes

+++ src/main/java/sonar/logistics/info/providers/tile/AE2EnergyProvider.java
@@ -31,7 +31,7 @@ public class AE2EnergyProvider extends TileProvider {
 
 	@Override
 	public void getHelperInfo(List<Info> infoList, World world, int x, int y, int z, ForgeDirection dir) {
-		byte id = this.getID();
+		int id = this.getID();
 		TileEntity target = world.getTileEntity(x, y, z);
 		if (target instanceof IAEPowerStorage) {
 			IAEPowerStorage power = (IAEPowerStorage) target;
@@ -49,12 +49,12 @@ public class AE2EnergyProvider extends TileProvider {
 	}
 
 	@Override
-	public String getCategory(byte id) {
+	public String getCategory(int id) {
 		return categories[id];
 	}
 
 	@Override
-	public String getSubCategory(byte id) {
+	public String getSubCategory(int id) {
 		return subcategories[id];
 	}
