+++ src/main/java/sonar/logistics/info/providers/tile/IC2ReactorProvider.java
@@ -32,7 +32,7 @@ public class IC2ReactorProvider extends TileProvider {
 
 	@Override
 	public void getHelperInfo(List<Info> infoList, World world, int x, int y, int z, ForgeDirection dir) {
-		byte id = this.getID();
+		int id = this.getID();
 		TileEntity target = world.getTileEntity(x, y, z);
 		IReactor reactor = null;
 		if (target instanceof IReactorChamber) {
@@ -52,12 +52,12 @@ public class IC2ReactorProvider extends TileProvider {
 	}
 
 	@Override
-	public String getCategory(byte id) {
+	public String getCategory(int id) {
 		return categories[id];
 	}
 
 	@Override
-	public String getSubCategory(byte id) {
+	public String getSubCategory(int id) {
 		return subcategories[id];
 	}
