+++ src/main/java/sonar/logistics/info/providers/tile/RFEnergyProvider.java
@@ -33,7 +33,7 @@ public class RFEnergyProvider extends TileProvider {
 
 	@Override
 	public void getHelperInfo(List<Info> infoList, World world, int x, int y, int z, ForgeDirection dir) {
-		byte id = this.getID();
+		int id = this.getID();
 		TileEntity handler = world.getTileEntity(x, y, z);
 		boolean displayEnergy = true;
 		if (Loader.isModLoaded("Mekanism")) {
@@ -64,12 +64,12 @@ public class RFEnergyProvider extends TileProvider {
 	}
 
 	@Override
-	public String getCategory(byte id) {
+	public String getCategory(int id) {
 		return categories[id];
 	}
 
 	@Override
-	public String getSubCategory(byte id) {
+	public String getSubCategory(int id) {
 		return subcategories[id];
 	}
 }
