+++ src/main/java/sonar/logistics/info/providers/tile/RotaryCraftProvider.java
@@ -35,7 +35,7 @@ public class RotaryCraftProvider extends TileProvider {
 
 	@Override
 	public void getHelperInfo(List<Info> infoList, World world, int x, int y, int z, ForgeDirection dir) {
-		byte id = this.getID();
+		int id = this.getID();
 
 		TileEntity handler = world.getTileEntity(x, y, z);
 		if (handler instanceof TemperatureTile) {
@@ -63,12 +63,12 @@ public class RotaryCraftProvider extends TileProvider {
 	}
 
 	@Override
-	public String getCategory(byte id) {
+	public String getCategory(int id) {
 		return categories[id];
 	}
 
 	@Override
-	public String getSubCategory(byte id) {
+	public String getSubCategory(int id) {
 		return subcategories[id];
 	}
